backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.49
botocore==1.42.49
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
firebase-admin==6.6.0
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.1.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
